Work on instructions and token resolving

+++ UScript.Compiler/Writer/Metadata/IMetadataToken.cs
@@ -4,7 +4,7 @@ namespace UScript.Compiler.Writer.Metadata
 {
     public interface IMetadataToken : IWriteable
     {
-        long TokenBase { get; }
-        long RawToken { get; set; }
+        int TokenBase { get; }
+        int RawToken { get; set; }
     }
 }
