+++ UScript.Compiler/Writer/Metadata/V1/FunctionMetadataToken.cs
@@ -5,11 +5,11 @@ namespace UScript.Compiler.Writer.Metadata.V1
 {
     public sealed class FunctionMetadataToken : IMetadataToken
     {
-        public long TokenBase => 0x10000000;
-        public long RawToken { get; set; }
+        public int TokenBase => 0x10000000;
+        public int RawToken { get; set; }
         public void WriteTo(Stream dest)
         {
-            dest.Write(BitConverter.GetBytes(RawToken), 0, sizeof(long));
+            dest.Write(BitConverter.GetBytes(RawToken), 0, sizeof(int));
         }
 
         public override string ToString()
