+++ UScript.Compiler/Writer/Metadata/V1/StringMetadataToken.cs
@@ -5,12 +5,12 @@ namespace UScript.Compiler.Writer.Metadata.V1
 {
     public sealed class StringMetadataToken : IMetadataToken
     {
-        public long TokenBase => 0x20000000;
-        public long RawToken { get; set; }
+        public int TokenBase => 0x20000000;
+        public int RawToken { get; set; }
 
         public void WriteTo(Stream dest)
         {
-            dest.Write(BitConverter.GetBytes(RawToken), 0, sizeof(long));
+            dest.Write(BitConverter.GetBytes(RawToken), 0, sizeof(int));
         }
 
         public override string ToString()
